Create plugin-api crate and move most of core's logic into it
