+++ Cargo.toml
@@ -3,6 +3,7 @@ members = [
   "crates/javy",
   "crates/core",
   "crates/cli",
+  "crates/plugin-api",
   "crates/test-macros",
   "crates/runner",
   "fuzz",
@@ -21,8 +22,8 @@ wasmtime = "19"
 wasmtime-wasi = "19"
 wasi-common = "19"
 anyhow = "1.0"
-once_cell = "1.20"
 javy = { path = "crates/javy", version = "3.0.2-alpha.1" }
+javy-plugin-api = { path = "crates/plugin-api", version = "1.0.0-alpha.1" }
 tempfile = "3.13.0"
 uuid = { version = "1.10", features = ["v4"] }
 serde = { version = "1.0", default-features = false }
