+++ crates/plugin-api/publish.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+set -e
+
+cargo publish --target=wasm32-wasip1
